chore: address quality-scorecard issues (#585, #586, #587, #588)
- fix(ic): short-circuit constant/degenerate windows in _engine_ic so
  np.corrcoef and scipy.spearmanr no longer emit warnings; constant inputs
  deterministically yield nan, with a covering regression test (#588)
- refactor(math): extract per-branch helpers so every function/method in
  src/ is radon grade A (no B-grade blocks remain); the dual-path
  maintenance note in _iter_solve is preserved (#587)
- refactor(math): split optimizer.py 786 -> 446 LOC into cohesive private
  modules (_engine_validation, _engine_core, _engine_performance,
  _engine_solve_base); the public/importable API is unchanged (#586)
- build: vendor a behaviour-aware test-layout checker (skips private
  modules, exempts cross-cutting suites) and wire `make test-layout` into
  the CI test job (#585)

Verified: 709 tests pass at 100% coverage, radon cc/mi all grade A,
ty + mypy strict clean, deptry clean, pre-commit (ruff/bandit/interrogate)
clean.
